// src/app/generated/api/model/item.ts
/**
 * Listok API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemStatus } from './itemStatus';
import { Product } from './product';


/**
 * An item
 */
export interface Item { 
    product: Product;
    /**
     * Quantity of the Item to be purchased.
     */
    count: number;
    /**
     * Optional color associated with the Item.
     */
    groupColor?: string;
    status: ItemStatus;
    /**
     * Unique identifier for the Item.
     */
    readonly id: string;
}